Move use_sdf_bounded_source tests to iobase.
In particular, these tests don't need to be run in every subclass of
FnApiRunnerTest.
